+++ DarmuhsTerminalCommands/Plugin.cs
@@ -9,7 +9,7 @@ using UnityEngine.UI;
 
 namespace TerminalStuff
 {
-    [BepInPlugin("darmuh.TerminalStuff", "darmuhsTerminalStuff", "3.0.3")]
+    [BepInPlugin("darmuh.TerminalStuff", "darmuhsTerminalStuff", "3.0.4")]
     [BepInDependency("atomic.terminalapi")]
     //[BepInDependency("Rozebud.FovAdjust")]
 
@@ -20,7 +20,7 @@ namespace TerminalStuff
         {
             public const string PLUGIN_GUID = "darmuh.lethalcompany.darmuhsTerminalStuff";
             public const string PLUGIN_NAME = "darmuhsTerminalStuff";
-            public const string PLUGIN_VERSION = "3.0.3";
+            public const string PLUGIN_VERSION = "3.0.4";
         }
 
         internal static new ManualLogSource Log;
@@ -33,6 +33,7 @@ namespace TerminalStuff
         public bool SolosBodyCamsMod = false;
         public bool OpenBodyCamsMod = false;
         public bool TwoRadarMapsMod = false;
+        public bool suitsTerminal = false;
 
         //public stuff for instance
         public bool awaitingConfirmation = false;
@@ -70,7 +71,7 @@ namespace TerminalStuff
         {
             Plugin.instance = this;
             Plugin.Log = base.Logger;
-            Plugin.Log.LogInfo((object)"Plugin darmuhsTerminalCommands is loaded with version 3.0.2!");
+            Plugin.Log.LogInfo((object)"Plugin darmuhsTerminalCommands is loaded with version 3.0.4!");
             Plugin.Log.LogInfo((object)"--------[Completely reworked for Optimal Compatibility]---------");
             ConfigSettings.BindConfigSettings();
             Harmony.CreateAndPatchAll(Assembly.GetExecutingAssembly());
@@ -127,13 +128,13 @@ namespace TerminalStuff
             AddKeywordIfEnabled(ConfigSettings.terminalLights.Value, TerminalEvents.AddLights);
             AddKeywordIfEnabled(ConfigSettings.terminalRandomSuit.Value, TerminalEvents.AddRandomSuit);
             AddKeywordIfEnabled(ConfigSettings.terminalClockCommand.Value, TerminalEvents.AddClockKeywords);
-            AddKeywordIfEnabled(ConfigSettings.terminalLobby.Value, TerminalEvents.AddCommandAction("lobby command\n", false, MoreCommands.otherActionNodes, "lobby", true, "Lobby Name", "lobby name", "", "", MoreCommands.GetLobbyName));
-            AddKeywordIfEnabled(ConfigSettings.terminalListItems.Value, TerminalEvents.AddCommandAction("List Items TermEvent\n", true, MoreCommands.otherActionNodes, ConfigSettings.ListItemsKeyword.Value, true, "List Items on Ship", "itemlist", "", "", MoreCommands.GetItemsOnShip));
-            AddKeywordIfEnabled(ConfigSettings.terminalListItems.Value, TerminalEvents.AddCommandAction("Detailed Loot TermEvent\n", true, MoreCommands.otherActionNodes, ConfigSettings.ListScrapKeyword.Value, true, "List Scrap on Ship", "lootlist", "", "", AllTheLootStuff.DetailedLootCommand));
-            AddKeywordIfEnabled(ConfigSettings.terminalBioScan.Value, TerminalEvents.AddCommandAction("bioscan terminal event\n", true, MoreCommands.infoOnlyNodes, "bioscan", true, "BioScan", "", "", CostCommands.BioscanCommand), ConfigSettings.ModNetworking.Value);
-            AddKeywordIfEnabled(ConfigSettings.terminalMirror.Value, TerminalEvents.AddCommandAction("mirror terminal event\n", true, ViewCommands.termViewNodes, "mirror", true, "Mirror", "", "", ViewCommands.MirrorEvent));
-            AddKeywordIfEnabled(ConfigSettings.terminalRefund.Value, TerminalEvents.AddCommandAction("refund terminal event\n", true, MoreCommands.infoOnlyNodes, "refund", true, "Refund", "", "", CostCommands.GetRefund), ConfigSettings.ModNetworking.Value); //unable to sync between clients without netpatch
-            AddKeywordIfEnabled(ConfigSettings.terminalPrevious.Value, TerminalEvents.AddCommandAction("switch back\n", true, ViewCommands.termViewNodes, "previous", true, "Switch to Previous", "", "", ViewCommands.HandlePreviousSwitchEvent));
+            AddKeywordIfEnabled(ConfigSettings.terminalLobby.Value, TerminalEvents.AddCommandAction("lobby command\n", false, MoreCommands.otherActionNodes, "lobby", false, "Lobby Name", "lobby name", "", "", MoreCommands.GetLobbyName));
+            AddKeywordIfEnabled(ConfigSettings.terminalListItems.Value, TerminalEvents.AddCommandAction("List Items TermEvent\n", true, MoreCommands.otherActionNodes, ConfigSettings.ListItemsKeyword.Value, false, "List Items on Ship", "itemlist", "", "", MoreCommands.GetItemsOnShip));
+            AddKeywordIfEnabled(ConfigSettings.terminalListItems.Value, TerminalEvents.AddCommandAction("Detailed Loot TermEvent\n", true, MoreCommands.otherActionNodes, ConfigSettings.ListScrapKeyword.Value, false, "List Scrap on Ship", "lootlist", "", "", AllTheLootStuff.DetailedLootCommand));
+            AddKeywordIfEnabled(ConfigSettings.terminalBioScan.Value, TerminalEvents.AddCommandAction("bioscan terminal event\n", true, MoreCommands.infoOnlyNodes, "bioscan", false, "BioScan", "", "", CostCommands.BioscanCommand), ConfigSettings.ModNetworking.Value);
+            AddKeywordIfEnabled(ConfigSettings.terminalMirror.Value, TerminalEvents.AddCommandAction("mirror terminal event\n", true, ViewCommands.termViewNodes, "mirror", false, "Mirror", "", "", ViewCommands.MirrorEvent));
+            AddKeywordIfEnabled(ConfigSettings.terminalRefund.Value, TerminalEvents.AddCommandAction("refund terminal event\n", true, MoreCommands.infoOnlyNodes, "refund", false, "Refund", "", "", CostCommands.GetRefund), ConfigSettings.ModNetworking.Value); //unable to sync between clients without netpatch
+            AddKeywordIfEnabled(ConfigSettings.terminalPrevious.Value, TerminalEvents.AddCommandAction("switch back\n", true, ViewCommands.termViewNodes, "previous", false, "Switch to Previous", "", "", ViewCommands.HandlePreviousSwitchEvent));
         }
 
         private static void AddKeywordIfEnabled(bool isEnabled, Action keywordAction)